recipe cards now working in DOM
