+++ src/domUpdates.js
@@ -1,16 +1,69 @@
-let headerUpdates = {
-  //headerDOMfunction
+let domUpdates = {
+//onLoad
+// emptyArea(name) {
+//   document.querySelector(name).innerHTML = '';
+// }
+//
+// findProperty(htmlTag, cssTag) {
+//   document.querySelector(htmlTag).classList.contains(cssTag)
+// }
+//
+// removeProperty(htmlTag, cssTag) {
+//   document.querySelector(htmlTag).classList.remove(cssTag)
+// }
 
-  //headerDOMFunction
+populateCards(recipes, user) {
+  let cardArea = document.querySelector('.all-cards');
+
+    cardArea.innerHTML = '';
+    if (cardArea.classList.contains('all')) {
+      cardArea.classList.remove('all')
+    }
+    recipes.forEach(recipe => {
+      cardArea.insertAdjacentHTML('afterbegin', `<div id='${recipe.id}'
+      class='card'>
+          <header id='${recipe.id}' class='card-header'>
+            <label for='add-button' class='hidden'>Click to add recipe</label>
+            <button id='${recipe.id}' aria-label='add-button' class='add-button card-button'>
+              <img id='${recipe.id} favorite' class='add'
+              src='https://image.flaticon.com/icons/svg/32/32339.svg' alt='Add to
+              recipes to cook'>
+            </button>
+            <label for='favorite-button' class='hidden'>Click to favorite recipe
+            </label>
+            <button id='${recipe.id}' aria-label='favorite-button' class='favorite favorite${recipe.id} card-button'></button>
+          </header>
+            <span id='${recipe.id}' class='recipe-name'>${recipe.name}</span>
+            <img id='${recipe.id}' tabindex='0' class='card-picture'
+            src='${recipe.image}' alt='click to view recipe for ${recipe.name}'>
+      </div>`)
+    })
+    this.getFavorites(user);
+},
+
+  greetUser(user) {
+    const userName = document.querySelector('.user-name');
+    userName.innerHTML =
+    user.name.split(' ')[0] + ' ' + user.name.split(' ')[1][0];
+  },
+
+  getFavorites(user) {
+    if (user.favoriteRecipes.length) {
+      user.favoriteRecipes.forEach(recipe => {
+        document.querySelector(`.favorite${recipe.id}`).classList.add('favorite-active')
+      })
+    } else return
+  }
+
+
+
+
+
+
 };
 
-// let mainUpdates = {
-//   changeMainView() {
-//
-//   }
-// };
 
 
 
 
-export default headerUpdates;
+export default domUpdates;
