+++ src/index.html
@@ -13,14 +13,14 @@
       <button id='home-button' class='home navButton'>Home</button>
       <label for='view-favorites-button' class='hidden'>View favorites</label>
       <button id='view-favorites-button' class=' view-favorites navButton'>View Favorites</button>
-      <!-- <form class='search-bar'>
+      <form class='search-bar'>
         <label for='search-button' class='hidden'>Click here to search here</label>
         <button id='search-button' class='search-button'>
           <img class='search-image' src='https://image.flaticon.com/icons/svg/660/660029.svg' alt='Search button'>
         </button>
         <label for='search-input' class='hidden'>Type search query here</label>
         <input id='search-input' class='search-input' type='text' placeholder='Search recipes by name or ingredients' value=''>
-      </form> -->
+      </form>
     </header>
     <main>
       <section class='all-cards'>
